pyproject.toml
[project]
name = "llms-performance"
version = "0.1.0"
description = ""
authors = [
    {name = "PhamVuThuNguyet",email = "nguyet.pham.researcher@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai (>=1.109.0,<2.0.0)",
    "anthropic (>=0.68.0,<0.69.0)",
    "google-genai (>=1.38.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "colorama (>=0.4.6,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
